==> 04_Figures/Figure4/06-R_ATAC_All_CD8.ipynb <==
quiet_library <- function(...) {
    suppressPackageStartupMessages(library(...))
}
quiet_library(Seurat)
quiet_library(ggplot2)
quiet_library(Matrix)
quiet_library(H5weaver)
quiet_library(dplyr)
quiet_library(viridis)
quiet_library(harmony)
quiet_library(Nebulosa)

quiet_library(ArchR)

addArchRThreads(32)
addArchRGenome("hg38")

# Load Data

sp_merge <- readRDS(file = '../../Seurat_Objects/sp_merge_gating_celltype_updated.rds')

sp_merge <- SetIdent(sp_merge, value = 'gating_celltype')
cd8_mem <- subset(sp_merge, idents = c('CD8 CM','CD8 EM1','CD8 EM2','CD8 TEMRA'))

cd8_mem$pure_celltype <- cd8_mem$gating_celltype

cd8na <- readRDS(file = 'trim_cd8_na.rds')

cd8_merge <- merge(cd8na, cd8_mem)

rm(sp_merge)
rm(cd8_mem)
rm(cd8na)

# ATAC

proj <- loadArchRProject(path = '../../PedSen_ATAC/')
proj

bcs <- colnames(cd8_merge[['RNA']])
head(bcs)

idx_new <- which(proj$cellNames_clean %in% bcs)
head(idx_new)

cells_subset <- proj$cellNames[idx_new]
proj_subset <- proj[cells_subset,]
proj_subset

### Add Pure Celltypes

pure_clusters <- as.data.frame(cd8_merge$pure_celltype)
rownames(pure_clusters) <- paste(cd8_merge$batch_id, "-P1_",cd8_merge$pbmc_sample_id,"#",rownames(pure_clusters), sep = "")
colnames(pure_clusters) <- 'pure_clusters'
pure_clusters[] <- lapply(pure_clusters, as.character)
head(pure_clusters)

proj_subset <- addCellColData(proj_subset, data = pure_clusters$pure_clusters, name = 'pure_celltype', cells = rownames(pure_clusters), force = TRUE)

# ATAC Clustering

proj_subset <- addIterativeLSI(proj_subset, name = 'IterativeLSI', force = TRUE)

proj_subset <- addHarmony(proj_subset, reducedDims = 'IterativeLSI', groupBy = 'batch_id', force = TRUE)

#proj_subset <- addClusters(proj_subset, reducedDims = 'IterativeLSI', name = 'Clusters', force = TRUE, resolution = 0.2)
proj_subset <- addClusters(proj_subset, reducedDims = 'Harmony', name = 'Clusters', force = TRUE, resolution = 0.2)

#proj_subset <- addUMAP(proj_subset, reducedDims = 'IterativeLSI', name = 'UMAP', force = TRUE)
proj_subset <- addUMAP(proj_subset, reducedDims = 'Harmony', name = 'UMAP', force = TRUE)

options(repr.plocd8_naidth = 8, repr.plot.height = 8)
plotEmbedding(ArchRProj = proj_subset, colorBy = "cellColData", name = "Clusters", embedding = "UMAP")
plotEmbedding(ArchRProj = proj_subset, colorBy = "cellColData", name = "pediatric_senior", embedding = "UMAP")
# plotEmbedding(ArchRProj = proj_subset, colorBy = "cellColData", name = "wnn_celltype", embedding = "UMAP")
plotEmbedding(ArchRProj = proj_subset, colorBy = "cellColData", name = "pure_celltype", embedding = "UMAP")

# Peak Analysis

proj_subset <- addGroupCoverages(ArchRProj = proj_subset, groupBy = "pure_celltype", maxReplicates = 16, maxCells = 2000, force = TRUE)

pathToMacs2 <- findMacs2()

proj_subset <- addReproduciblePeakSet(
                        ArchRProj = proj_subset, 
                        groupBy = "pure_celltype", 
                        pathToMacs2 = pathToMacs2, force = TRUE
                            )

proj_subset <- addPeakMatrix(proj_subset)

# peak heatmap

table(proj_subset$pure_celltype)

markersPeaks_pure <- getMarkerFeatures(maxCells = 1000,
    ArchRProj = proj_subset,
    useMatrix = "PeakMatrix", 
    groupBy = "pure_celltype",
  bias = c("TSSEnrichment", "log10(nFrags)"),
  testMethod = "wilcoxon"
)

heatmapPeaks_pure <- markerHeatmap(
  seMarker = markersPeaks_pure, 
  cutOff = "FDR <= 0.1 & Log2FC >= 0.5",
  transpose = FALSE
)

options(repr.plot.width = 8, repr.plot.height = 12)
draw(heatmapPeaks_pure, heatmap_legend_side = "right", annotation_legend_side = "bot")

# Open a pdf file
pdf("plots/cd8all_peaks_atac_heatmap.pdf", width = 8, height = 12) 
# 2. Create a plot
draw(heatmapPeaks_pure, heatmap_legend_side = "right", annotation_legend_side = "bot")
# Close the pdf file
dev.off()
